moved pie chart into its own component file
